+++ .gitignore
@@ -17,6 +17,9 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.eslintcache
+
+Dockerfile
 
 npm-debug.log*
 yarn-debug.log*
