+++ package.json
@@ -2,11 +2,14 @@
   "name": "collect-and-search-ui",
   "version": "0.1.0",
   "private": true,
+  "homepage": "http://netcracker-education-center.github.io/collect-and-search-ui",
   "dependencies": {
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.0",
     "bootstrap": "^4.6.0",
+    "create-react-context": "^0.3.0",
+    "gh-pages": "^3.1.0",
     "jquerry": "0.0.1-security",
     "jquery": "^3.5.1",
     "multiselect-react-dropdown": "^1.6.7",
@@ -25,7 +28,9 @@
     "start": "set PORT=8080 && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build"
   },
   "eslintConfig": {
     "extends": [
