+++ src/App.js
@@ -3,7 +3,6 @@ import React from 'react';
 import Header from './Components/Header';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
-//import Modal from './Components/Modal';
 
 const App = () => {
