+++ src/Pages/Home.js
@@ -10,7 +10,6 @@ export default class Home extends Component {
             issues: [],
             shawIssues: false
         }
-        //this.newRequest();
     }
 
     newRequest() {
@@ -18,17 +17,14 @@ export default class Home extends Component {
 
         var data;
         axios.post("http://localhost:9090/tickets/find", {
-            //body: {
             ticketSystem: "JIRA",
             login: "kakashka_am@mail.ru",
             password: 'IdBigXbJL2aIgrJhGGg2B1A8',
             url: "https://netcrackereducation.atlassian.net/"
-            //}
-            //accept: "application/json"
         }).then(response => {
             data = response.data;
             console.log('data: ', data);
-            this.setState(state => {return{ issues: data, shawIssues: true }});
+            this.setState(() => {return{ issues: data, shawIssues: true }});
         }).catch(error => {
             console.log("request error: ", error)
         });
