+++ src/index.js
@@ -1,10 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/js/bootstrap.bundle.min';
-//import './index.css';
 import App from './App';
-//import Field from './Field';
-//import Modal from './Components/Modal';
 import reportWebVitals from './reportWebVitals';
